4.2. Результат работы метода [#232622]

+++ src/ru/job4j/calculator/MathFunc.java
@@ -1,7 +1,7 @@
 package ru.job4j.calculator;
 
 public class MathFunc {
-    public static float func1(float x) {
+    public static int func1(int x) {
         return x * x + 1;
     }
 
@@ -10,9 +10,9 @@ public class MathFunc {
     }
 
     public static void main(String[] args) {
-        float result1 = MathFunc.func1(3);
+        int result1 = MathFunc.func1(3);
         float result2 = MathFunc.func2(5);
-        float result3 = MathFunc.func1(100);
+        int result3 = MathFunc.func1(100);
         float total = result1 + result2;
         System.out.println(total);
         System.out.println(result3);
